separate out makeTimingTable into Util.Timing module.
